Attempt to promote variables to function arguments
This turns out to be too tricky to do in a safe way.  Our attempt is to
detect when variables are always used initialized; in this case we can
safely promote them to function args (well assuming eval/with is not used
and the function is either `strict` or doesn't use `arguments`).  Saves
quite a few bytes on my code, but it's still unsafe.

I'm just dropping my attempt in this branch and give up.